+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+## 0.11.3
+* Feature
+  - add: `props.footerConfig.position` which can be set footer to top. 
+  - Add: `props.sortConfig.multiSort`
+  - Add: `props.sortConfig.multiSortLimit`
+
 ## 0.11.2
 * Bugfix:
   - fix: areaSelection keyboard left right not work correctly when table has fixed sticky column